backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
supabase
python-multipart
Pillow
pypdf
python-jose[cryptography]
passlib[bcrypt]
slowapi
psycopg2-binary
